create a new empty tab for the asset details
